Slim down distinctpy stubs, sort imports
