+++ docs/ref/cli.py
@@ -1,6 +1,7 @@
-from typer.cli import app
 from contextlib import redirect_stdout
+
 import mkdocs_gen_files
+from typer.cli import app
 
 with mkdocs_gen_files.open("ref/cli.md", "w") as f:
     with redirect_stdout(f):
