+++ python/evalio/__init__.py
@@ -1,9 +1,10 @@
-from . import _cpp, datasets, pipelines, types, utils, stats
-from ._cpp import abi_tag as _abi_tag
-
-import warnings
-from tqdm import TqdmExperimentalWarning
 import atexit
+import warnings
+
+from tqdm import TqdmExperimentalWarning
+
+from . import _cpp, datasets, pipelines, stats, types, utils
+from ._cpp import abi_tag as _abi_tag
 
 Param = bool | int | float | str
