+++ python/evalio/datasets/__init__.py
@@ -1,15 +1,14 @@
-from .base import Dataset, DatasetIterator, set_data_dir, get_data_dir
+from .base import Dataset, DatasetIterator, get_data_dir, set_data_dir
 from .botanic_garden import BotanicGarden
 from .cumulti import CUMulti
 from .enwide import EnWide
 from .helipr import HeLiPR
 from .hilti_2022 import Hilti2022
+from .loaders import RawDataIter, RosbagIter
+from .multi_campus import MultiCampus
 from .newer_college_2020 import NewerCollege2020
 from .newer_college_2021 import NewerCollege2021
-from .multi_campus import MultiCampus
 from .oxford_spires import OxfordSpires
-
-from .loaders import RawDataIter, RosbagIter
 
 __all__ = [
     "get_data_dir",
