+++ python/evalio/datasets/multi_campus.py
@@ -1,4 +1,9 @@
+import os
 from enum import auto
+from pathlib import Path
+from typing import Optional, Sequence
+
+import numpy as np
 
 from evalio.datasets.loaders import (
     LidarDensity,
@@ -8,18 +13,12 @@ from evalio.datasets.loaders import (
     LidarStamp,
     RosbagIter,
 )
-from evalio.types import Trajectory, SE3, ImuParams, LidarParams
-import numpy as np
-
-import os
+from evalio.types import SE3, ImuParams, LidarParams, Trajectory
 
 from .base import (
     Dataset,
     DatasetIterator,
 )
-
-from pathlib import Path
-from typing import Sequence, Optional
 
 
 class MultiCampus(Dataset):
