+++ python/evalio/datasets/newer_college_2021.py
@@ -1,3 +1,10 @@
+import os
+from enum import auto
+from pathlib import Path
+from typing import Optional, Sequence
+
+import numpy as np
+
 from evalio.datasets.loaders import (
     LidarDensity,
     LidarFormatParams,
@@ -6,18 +13,12 @@ from evalio.datasets.loaders import (
     LidarStamp,
     RosbagIter,
 )
-from evalio.types import Trajectory, SO3, SE3, ImuParams, LidarParams
-import numpy as np
-from enum import auto
+from evalio.types import SE3, SO3, ImuParams, LidarParams, Trajectory
+
 from .base import (
     Dataset,
     DatasetIterator,
 )
-
-import os
-
-from pathlib import Path
-from typing import Sequence, Optional
 
 
 class NewerCollege2021(Dataset):
