+++ python/evalio/types.py
@@ -1,23 +1,22 @@
+import csv
+from dataclasses import dataclass, field
+from pathlib import Path
+from typing import Optional
+
+import numpy as np
+import yaml
+
 from ._cpp.types import (  # type: ignore
     SE3,
     SO3,
     Duration,
-    Stamp,
-    Point,
+    ImuMeasurement,
+    ImuParams,
     LidarMeasurement,
     LidarParams,
-    ImuParams,
-    ImuMeasurement,
+    Point,
+    Stamp,
 )
-from dataclasses import dataclass, field
-
-from pathlib import Path
-import yaml
-
-from typing import Optional
-import csv
-
-import numpy as np
 
 
 @dataclass(kw_only=True)
