+++ python/typings/distinctipy/__init__.pyi
@@ -2,36 +2,71 @@
 This type stub file was generated by pyright.
 """
 
-from .distinctipy import (
-    BLACK,
-    BLUE,
-    CORNERS,
-    CYAN,
-    GREEN,
-    INTERIOR,
-    MAGENTA,
-    MID_FACE,
-    POINTS_OF_INTEREST,
-    RED,
-    WHITE,
-    YELLOW,
-    get_colors,
-)
+WHITE = ...
+BLACK = ...
+RED = ...
+GREEN = ...
+BLUE = ...
+CYAN = ...
+YELLOW = ...
+MAGENTA = ...
+CORNERS = ...
+MID_FACE = ...
+INTERIOR = ...
+POINTS_OF_INTEREST = ...
+_SEED_MAX = ...
 
-"""
-distinctipy is a lightweight python package providing functions to generate
-colours that are visually distinct from one another.
+def get_colors(
+    n_colors: int,
+    exclude_colors: list[tuple[float, float, float]] = ...,
+    return_excluded: bool = ...,
+    pastel_factor: float = ...,
+    n_attempts: int = ...,
+    colorblind_type: str = ...,
+    rng: int = ...,
+) -> list[tuple[float, float, float]]:
+    """
+    Generate a list of n visually distinct colours.
 
-Example:
-    >>> import distinctipy
-    >>> distinctipy.get_colors(5)
-    [(0, 1, 0), (1, 0, 1), (0, 0.5, 1), (1, 0.5, 0), (0.5, 0.75, 0.5)]
-"""
-name = ...
-__version__ = ...
-__external__ = ...
-__protected__ = ...
-__autogen_notes__ = ...
+    :param n_colors: How many colours to generate
+
+    :param exclude_colors: A list of (r,g,b) colours that new colours should be distinct
+        from. If exclude_colours=None then exclude_colours will be set to avoid white
+        and black (exclude_colours=[(0,0,0), (1,1,1)]). (r,g,b) values should be floats
+        between 0 and 1.
+
+    :param return_excluded: If return_excluded=True then exclude_colors will be included
+        in the returned color list. Otherwise only the newly generated colors are
+        returned (default).
+
+    :param pastel_factor: float between 0 and 1. If pastel_factor>0 paler colours will
+        be generated.
+
+    :param n_attempts: number of random colours to generated to find most distinct
+        colour.
+
+    :param colorblind_type: Generate colours that are distinct with given type of
+        colourblindness. Can be:
+
+            * 'Normal': Normal vision
+            * 'Protanopia': Red-green colorblindness (1% males)
+            * 'Protanomaly': Red-green colorblindness (1% males, 0.01% females)
+            * 'Deuteranopia': Red-green colorblindness (1% males)
+            * 'Deuteranomaly': Red-green colorblindness (most common type: 6% males,
+            0.4% females)
+            * 'Tritanopia': Blue-yellow colourblindness (<1% males and females)
+            * 'Tritanomaly' Blue-yellow colourblindness (0.01% males and females)
+            * 'Achromatopsia': Total colourblindness
+            * 'Achromatomaly': Total colourblindness
+
+    :param rng: A random integer seed or random.Random state.
+        If unspecified the global random is used.
+
+    :return: colors - A list of (r,g,b) colors that are visually distinct to each other
+        and to the colours in exclude_colors. (r,g,b) values are floats between 0 and 1.
+    """
+    ...
+
 __all__ = [
     "BLACK",
     "BLUE",
@@ -46,5 +81,4 @@ __all__ = [
     "WHITE",
     "YELLOW",
     "get_colors",
-    "name",
 ]
